requirements.txt
pandas~=2.3.0
newspaper3k~=0.2.8
nltk~=3.9.1